+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.0.0-beta8] - In progress
+
+### Fixed
+
+- ManyToOne relationship with null value, do not alter entity
+
 ## [1.0.0-beta7] - 2022-06-24
 
 ### Changed